+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 11.2.0
+
+* [#264](https://github.com/GEOSwift/GEOSwift/pull/264) Add initial support for prepared geometry
+
+## 11.1.0
+
+* [#290](https://github.com/GEOSwift/GEOSwift/pull/290) Add `LineStringConvertible.substring(fromFraction:toFraction:)`
+
+## 11.0.0
+
+* [#284](https://github.com/GEOSwift/GEOSwift/pull/284) Update GEOS, CI, and Apple Platform Versions
+* [#285](https://github.com/GEOSwift/GEOSwift/pull/285) Fix crash in polygonize()
+
+## 10.3.0
+
+* [#281](https://github.com/GEOSwift/GEOSwift/pull/281) Add `GeometryConvertible.lineMergeDirected()` and `GeometryConvertible.lineMerge()`
+* [#282](https://github.com/GEOSwift/GEOSwift/pull/282) Add `GeometryConvertible.offsetCurve(width:quadsegs:joinStyle:mitreLimit:)`
+
 ## 10.2.0
 
 * [#271](https://github.com/GEOSwift/GEOSwift/pull/271) Expose `GeometryConvertible.bufferWithStyle(width:quadsegs:endCapStyle:joinStyle:mitreLimit:)`